http,html: format directory entries and URIs
This patch fixes two problems.  First, directory URIs were not always
forced to end in a '/', which caused problems in relative links.
Second, when rendering directory entries, the code did not distinguish
between sub-directories and files (and instead displayed them all in
alphabetical order).

This patch introduces a 302 redirect when the URI points to a directory
so that the URI always ends in a '/'.  This patch also sorts the
directory entries so that all sub-directories and displayed before
files.