config: swap flags for JSON configuration file (#6)
The set of options passed to the binary are about to increase
substantially, so instead of having a long list of cumbersome flags
(with some levels of nesting), this patch switches to using a JSON file
that contains the configuration.

This patch makes the configuration file mandatory, and this decision is
intentional.  Instead of leaving the end user to guess the default
values for all options, the project comes bundled with a configuration
file that explicitly lists the default values.