config: add elementary configuration for OAuthV2-based authentication
This patch adds some basic configuration fields useful for
authentication/authorization using OAuthV2.  Initially, I plan to
support just GitHub as the OAuth provider, although the list should be
easy to extend.

An an ancillary change, this patch drops the file patterns that gate the
GitHub workflow run, since the workflow is required to merge PRs, but
the workflow is not run for files that are not C/C++ sources.