config: initial validation of auth configuration and move to library
The configuration validation code is large enough that it needs its own
test cases, so this patch moves the configuration processing to the
library code and out of the main application.  This patch also adds
validation for the `auth` section of the configuration, in addition to
adding tests to validate the validation.