config: separate config.json from default configuration
This patch introduces a new file, ".default.config.json", which holds
the default configuration and tells Git to ignore the changes in the
actual configuration file, "config.json".  If the config.json file does
not exist, the contents of .default.config.json are copied over.  This
allows freely modifying the configuration file without having to stash
and unstash end user's changes while pulling the most recent commit.

As an ancillary change, this patch renames the `check-server` test
target to `check-magenta` and it moves the build rules for external
sources into the externals/ directory.